requirements.txt
fastapi
uvicorn[standard]
librosa
numpy
scikit-learn
python-dotenv
soundfile
joblib


